Updated sizeof integer
Microprocessor compiler converts int to a 16bit value, where as a 32bit int is needed

+++ Full.c
@@ -179,7 +179,7 @@ void SplitNum(int Num, int * Arr){
 
 }
 
-void GenCodes(int *Keys){
+void GenCodes(long int *Keys){
 	int a, b, c, d, e, f, counter;	
 	counter = 0;
